Add hots to games module
